-add support for existing @return iterable -refactoring
